bot/requirements.txt
requests==2.23.0
google-api-python-client
oauth2client==4.1.3 
Flask==1.1.2
